Refactor code and save data to json
